+++ src/loaders/iloader.ts
@@ -1,0 +1,3 @@
+import { ILoaderResult } from "./iloaderResult";
+
+export type ILoader = (scriptPath: string, options?: any) => Promise<ILoaderResult>;