+++ src/loaders/index.ts
@@ -1,0 +1,7 @@
+import { tsnodeLoader } from "./tsnodeLoader";
+import { typescriptLoader } from "./typescriptLoader";
+
+export const loaders = [
+    tsnodeLoader,
+    typescriptLoader,
+];